06-lab-complete/requirements.txt
streamlit==1.44.1
flask==3.1.0
openai==1.76.0
python-dotenv==1.0.1
pypdf==6.0.0
